+++ SpellActivationOverlay.lua
@@ -61,16 +61,19 @@ function SpellActivationOverlay_OnLoad(self)
 --	self:RegisterUnitEvent("UNIT_AURA", "player");
 	self:RegisterEvent("COMBAT_LOG_EVENT_UNFILTERED");
 	self:RegisterEvent("PLAYER_ENTERING_WORLD");
-	self:RegisterEvent("SPELL_UPDATE_USABLE");
 	self:RegisterEvent("PLAYER_REGEN_ENABLED");
 	self:RegisterEvent("PLAYER_REGEN_DISABLED");
 	self:RegisterEvent("SPELLS_CHANGED");
 	self:RegisterEvent("LEARNED_SPELL_IN_TAB");
 	self:RegisterEvent("LOADING_SCREEN_DISABLED");
 	self:RegisterEvent("PLAYER_LOGIN");
-	self:RegisterEvent("PLAYER_TALENT_UPDATE");
-	if ( SAO.IsCata() and classFile == "PALADIN" ) then
-		self:RegisterEvent("UNIT_POWER_FREQUENT"); -- For Holy Power, introduced in Cataclysm
+	for _, var in pairs(SAO.Variables) do
+		if type(var.event.isRequired) == 'function' and var.event.isRequired()
+		or type(var.event.isRequired) == 'boolean' and var.event.isRequired then
+			for _, eventName in ipairs(var.event.names) do
+				self:RegisterEvent(eventName);
+			end
+		end
 	end
 end