+++ options/defaults.lua
@@ -22,7 +22,7 @@ SAO.defaults = {
             },
             glow = {
                 [49222] = { -- Bone Shield (Cataclysm)
-                    [49222] = true, -- Bone Shield
+                    [49222] = false, -- Bone Shield
                 },
                 [56815] = { -- Rune Strike (Wrath)
                     [56815] = true, -- Rune Strike
@@ -334,6 +334,11 @@ SAO.defaults = {
                 [53600] = { -- Shield of the Righteous (Cataclysm)
                     [53600] = true, -- Shield of the Righteous
                 },
+                [53671] = { -- Judgements of the Pure (Wrath+)
+                    [20271] = true, -- Judgement of Light (Wrath) / Judgement (Cata)
+                    [53408] = true, -- Judgement of Wisdom (Wrath only)
+                    [53407] = true, -- Judgement of Justice (Wrath only)
+                },
                 [407778] = { -- Divine Storm (Season of Discovery)
                     [407778]= true, -- Divine Storm (Season of Discovery)
                 },
@@ -408,6 +413,15 @@ SAO.defaults = {
                 },
             },
             glow = {
+                [588] = { -- Inner Fire (Wrath+)
+                    [588] = true, -- Inner Fire
+                },
+                [15473] = { -- Shadowform
+                    [15473] = true, -- Shadowform
+                },
+                [32379] = { -- Shadow Word: Death (Cataclysm)
+                    [32379] = true, -- Shadow Word: Death
+                },
                 [33151] = { -- Surge of Light (TBC - Wrath)
                     [585]  = true, -- Smite
                     [2061] = true, -- Flash Heal (not for TBC)
@@ -448,6 +462,9 @@ SAO.defaults = {
                 },
             },
             glow = {
+                [14158] = { -- Murderous Intent (not an actual buff) (Cataclysm)
+                    [53] = true, -- Backstab
+                },
                 [14251] = { -- Riposte (Era - Wrath)
                     [14251] = "cd:off", -- Riposte
                 },
@@ -552,10 +569,16 @@ SAO.defaults = {
                 [47283] = { -- Empowered Imp (Wrath+)
                     [0] = true,
                 },
+                [89937] = { -- Fel Spark (Cataclysm)
+                    [0] = true, -- any stacks
+                },
             },
             glow = {
                 [1120] = { -- Drain Soul (Wrath+)
                     [1120] = "spec:1", -- Drain Soul
+                },
+                [17877] = { -- Shadowburn
+                    [17877] = true, -- Shadowburn
                 },
                 [17941] = { -- Nightfall
                     [686] = true, -- Shadow Bolt
@@ -578,6 +601,9 @@ SAO.defaults = {
                 },
                 [47283] = { -- Empowered Imp (Cataclysm)
                     [6353]  = true, -- Soul Fire
+                },
+                [89937] = { -- Fel Spark (Cataclysm)
+                    [77799]  = true, -- Fel Flame
                 },
             },
         },